Add files via upload
changed the structure of the prediction module the main arithmetic works will now be done in the modes because the hardwaredistribution can not be taken for granted when calculating the powerconsumption and the time it takes the hardwareelements to finish the job. started to add comments. abstract class problem still exists